requirements.txt
scapy.all
colorama
argparse